Refactored to fit UML diagram.  Created new classes to fit UML diagram.
