+++ UrbanParks/UrbanParksMain.java
@@ -1,0 +1,18 @@
+/**
+ * Starts the program for the Urban Parks application.
+ * 
+ * TCSS 360, Winter 2016
+ * @author Group 6: Jonathan Hughes, Michael Ford, Weiwei Shi, Chris Vishoot
+ * @version January 21, 2016
+ */
+public class UrbanParksMain {
+
+    /**
+     * Main method for the Urban Parks user interface.
+     * @param args Command line args
+     */
+    public static void main(String[] args) {
+        GUI gui = new GUI(); //create GUI
+        gui.run();           //run GUI
+    }
+}
