+++ UrbanParks/User.java
@@ -1,0 +1,3 @@
+public abstract class User {
+    //TODO stuff
+}
