--- Urban_Parks/UrbanParksJob.java
@@ -1,3 +1,0 @@
-public class UrbanParksJob {
-    //TODO stuff
-}
