--- Urban_Parks/UrbanParksJobs.java
@@ -1,3 +1,0 @@
-public class UrbanParksJobs {
-    //TODO stuff
-}
