--- Urban_Parks/UrbanParksUser.java
@@ -1,3 +1,0 @@
-public abstract class UrbanParksUser {
-    //TODO stuff
-}
