--- Urban_Parks/UrbanParksUsers.java
@@ -1,3 +1,0 @@
-public class UrbanParksUsers {
-    //TODO stuff
-}
